Changing groupid to org.jboss.pnc.projectmanipulator
It's mostly done because any PNC admin can deploy project-manipulator from
groupid org/jboss/pnc/*
