--- core/src/main/java/org/jboss/projectmanipulator/core/ManipulationException.java
@@ -1,79 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2018-2020 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.projectmanipulator.core;
-
-import java.text.MessageFormat;
-
-public class ManipulationException extends Exception {
-
-    private static final long serialVersionUID = 1L;
-
-    private Object[] params;
-
-    private String formattedMessage;
-
-    public ManipulationException(final String messageFormat, final Throwable cause, final Object... params) {
-        super(messageFormat, cause);
-        this.params = params;
-    }
-
-    public ManipulationException(final String string, final Object... params) {
-        this(string, null, params);
-    }
-
-    @Override
-    public synchronized String getMessage() {
-        if (formattedMessage == null) {
-            final String format = super.getMessage();
-            if (params == null || params.length < 1) {
-                formattedMessage = format;
-            } else {
-                try {
-                    formattedMessage = String.format(format.replaceAll("\\{\\}", "%s"), params);
-                } catch (final Error | Exception e) {
-                }
-
-                if (formattedMessage == null || format == formattedMessage) {
-                    try {
-                        formattedMessage = MessageFormat.format(format, params);
-                    } catch (final Error | RuntimeException e) {
-                        formattedMessage = format;
-                        throw e;
-                    } catch (final Exception e) {
-                        formattedMessage = format;
-                    }
-                }
-            }
-        }
-
-        return formattedMessage;
-    }
-
-    private Object writeReplace() {
-        final Object[] newParams = new Object[params.length];
-        int i = 0;
-        for (final Object object : params) {
-            newParams[i] = String.valueOf(object);
-            i++;
-        }
-
-        this.params = newParams;
-        return this;
-    }
-
-}
