+++ npm/pom.xml
@@ -22,7 +22,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>org.jboss.project-manipulator</groupId>
+    <groupId>org.jboss.pnc.project-manipulator</groupId>
     <artifactId>project-manipulator-parent</artifactId>
     <version>1.0.2-SNAPSHOT</version>
   </parent>
@@ -33,7 +33,7 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.jboss.project-manipulator</groupId>
+      <groupId>org.jboss.pnc.project-manipulator</groupId>
       <artifactId>project-manipulator-core</artifactId>
     </dependency>
