+++ npm/src/main/java/org/jboss/pnc/projectmanipulator/npm/NpmResult.java
@@ -1,0 +1,72 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2018-2020 Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.pnc.projectmanipulator.npm;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+/**
+ * Represents result of manipulation.
+ */
+public class NpmResult {
+
+    /** Resulting package name. */
+    private String name;
+
+    /** Resulting package version. */
+    private String version;
+
+    /** Modified depedencies version. */
+    private Map<String, String> dependenciesMap = new LinkedHashMap<String, String>();
+
+    /** Modified devDepedencies version. */
+    private Map<String, String> devDependenciesMap = new LinkedHashMap<String, String>();
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    public Map<String, String> getDependenciesMap() {
+        return dependenciesMap;
+    }
+
+    public void setDependenciesMap(Map<String, String> dependenciesMap) {
+        this.dependenciesMap = dependenciesMap;
+    }
+
+    public Map<String, String> getDevDependenciesMap() {
+        return devDependenciesMap;
+    }
+
+    public void setDevDependenciesMap(Map<String, String> devDependenciesMap) {
+        this.devDependenciesMap = devDependenciesMap;
+    }
+
+}
