--- npm/src/main/java/org/jboss/projectmanipulator/npm/NpmManipulationSessionFactory.java
@@ -1,33 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2018-2020 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.projectmanipulator.npm;
-
-import java.io.File;
-import java.util.Properties;
-
-public class NpmManipulationSessionFactory {
-
-    public static NpmManipulationSession createSession(
-            File projectFile,
-            File resultFile,
-            Properties properties,
-            Properties userProps) {
-        return new NpmManipulationSession(projectFile, resultFile, properties, userProps);
-    }
-
-}
